+++ src/allowance/utils.ts
@@ -1,11 +1,17 @@
 import { TransactionRequest } from '@ethersproject/abstract-provider'
 import { ChainId, Token } from '@lifi/types'
 import BigNumber from 'bignumber.js'
-import { Contract, ContractTransaction, Signer } from 'ethers'
+import {
+  Contract,
+  ContractTransaction,
+  PopulatedTransaction,
+  Signer,
+  ethers,
+} from 'ethers'
 import ChainsService from '../services/ChainsService'
 import { ERC20Contract, ERC20_ABI, RevokeTokenData } from '../types'
 import { ServerError } from '../utils/errors'
-import { fetchDataUsingMulticall, MultiCallData } from '../utils/multicall'
+import { MultiCallData, fetchDataUsingMulticall } from '../utils/multicall'
 
 export const getApproved = async (
   signer: Signer,
@@ -29,15 +35,36 @@ export const getApproved = async (
   }
 }
 
-export const setApproval = (
+export const setApproval = async (
   signer: Signer,
   tokenAddress: string,
   contractAddress: string,
-  amount: string
-): Promise<ContractTransaction> => {
+  amount: string,
+  returnPopulatedTransaction?: boolean
+): Promise<ContractTransaction | PopulatedTransaction> => {
   const erc20 = new Contract(tokenAddress, ERC20_ABI, signer) as ERC20Contract
 
-  return erc20.approve(contractAddress, amount)
+  if (returnPopulatedTransaction) {
+    return erc20.populateTransaction.approve(contractAddress, amount)
+  }
+
+  const transactionRequest = await erc20.populateTransaction.approve(
+    contractAddress,
+    amount
+  )
+
+  try {
+    const estimatedGasLimit = await signer.estimateGas(transactionRequest)
+    if (estimatedGasLimit) {
+      const formattedGasLimit = ethers.BigNumber.from(
+        `${(BigInt(estimatedGasLimit.toString()) * 125n) / 100n}`
+      )
+
+      transactionRequest.gasLimit = formattedGasLimit
+    }
+  } catch (error) {}
+
+  return signer.sendTransaction(transactionRequest)
 }
 
 export const getAllowanceViaMulticall = async (